don't allow user to draw a new box until they save or cancel the previous one.  also disable scroll zooming

+++ js/src/annotations/osd-region-rect-tool.js
@@ -13,6 +13,7 @@
       rectClass:   null,
       osdOverlay:  null,
       dragging:    false,
+      rectangleDrawn: false,
       parent:      null
       }, options);
       
@@ -96,15 +97,17 @@
 
     startRectangle: function(event) {
       var _this = this.userData.recttool; //osd userData
-      if (!_this.dragging) {
-        _this.dragging = true; 
-        _this.mouseStart = _this.getMousePositionInImage(_this.osdViewer.viewport.pointFromPixel(event.position));
-        _this.createRectangle(_this.mouseStart);
-        _this.onDrawStart();
-      } else { 
-        var mouseNow = _this.getMousePositionInImage(_this.osdViewer.viewport.pointFromPixel(event.position));
-        _this.updateRectangle(_this.mouseStart, mouseNow);
-        _this.onDraw();
+      if (!_this.rectangleDrawn) {
+        if (!_this.dragging) {
+          _this.dragging = true; 
+          _this.mouseStart = _this.getMousePositionInImage(_this.osdViewer.viewport.pointFromPixel(event.position));
+          _this.createRectangle(_this.mouseStart);
+          _this.onDrawStart();
+        } else { 
+          var mouseNow = _this.getMousePositionInImage(_this.osdViewer.viewport.pointFromPixel(event.position));
+          _this.updateRectangle(_this.mouseStart, mouseNow);
+          _this.onDraw();
+        }
       }
     },
 
@@ -151,7 +154,7 @@
           width: Math.max(parseInt(osdImageRect.width, 10), 1),  //don't allow 0 pixel width or height
           height: Math.max(parseInt(osdImageRect.height, 10), 1) //don't allow 0 pixel width or height
         };
-
+        _this.rectangleDrawn = true;
         _this.onDrawFinish(canvasRect);
         _this.rectangle = null;
         _this.mouseStart = null;
@@ -237,7 +240,9 @@
               
                 //disable all tooltips for overlays
                 jQuery.publish('disableTooltips.'+parent.windowId);
+                //disable zooming
                 _this.osdViewer.zoomPerClick = 1;
+                _this.osdViewer.zoomPerScroll = 1;
 
                 tinymce.init({
                   selector : 'form.annotation-tooltip textarea',
@@ -260,11 +265,20 @@
               
                 jQuery('.annotation-tooltip a.cancel').on("click", function(event) {
                   event.preventDefault();
+                  //add check so that dialog box only pops up if there is stuff in the editor
+                  var content = tinymce.activeEditor.getContent();
+                  if (content) {
+                    if (!window.confirm("Do you want to cancel this annotation?")) { 
+                      return false;
+                    }
+                  }
                   api.destroy();
+                  _this.rectangleDrawn = false;
                   _this.osdViewer.removeOverlay(_this.osdOverlay);
                   //reenable viewer tooltips
                   jQuery.publish('enableTooltips.'+parent.windowId);
                   _this.osdViewer.zoomPerClick = 2;
+                  _this.osdViewer.zoomPerScroll = 1.2;                  
                 });
                 
                 jQuery('.annotation-tooltip a.save').on("click", function(event) {
@@ -324,9 +338,11 @@
 
                 //update content of this qtip to make it a viewer, not editor
                 api.destroy();
+                _this.rectangleDrawn = false;
                 //reenable viewer tooltips
                 jQuery.publish('enableTooltips.'+parent.windowId);
                 _this.osdViewer.zoomPerClick = 2;
+                _this.osdViewer.zoomPerScroll = 1.2;                  
                 });
               }
             }
